requirements.txt
python-telegram-bot==20.3
Telethon==1.30.0
pytz
